fix error in 32bit arch

+++ varint.go
@@ -14,7 +14,7 @@ func encodeVarint(value int32) []byte {
 func decodeVarint(input []uint8) int32 {
 	var ret int32
 	for i, value := range input {
-		ret |= (int32(value) & 127) << (7 * i)
+		ret |= (int32(value) & 127) << uint32(7*i)
 		if value&128 == 0 {
 			break
 		}
